+++ src/diag_cluster.h
@@ -33,7 +33,7 @@ class DiagCluster : public Diag {
   void dump_clusters(double);
   void dump_clusters_detailed();
   void generate_clusters();
-  void add_cluster(int, int, int, int*);
+  void add_cluster(int, double, int, double*);
   void free_clustlist();
 
   int* cluster_ids;
